added a transactions component and just put it on homepage to test
